optimizador/requirements.txt
opencv-python
pandas
numpy
matplotlib.pyplot

